add logger && refactor client
